+++ config/healthcheck.php
@@ -101,6 +101,11 @@ return [
         'ignore' => [],
     ],
 
+    'scheduler' => [
+        'timestamp-filename' => 'laravel-scheduler-health-check.txt',
+        'minutes_between_checks' => 5,
+    ],
+
     /**
      * Additional config can be put here. For example, a health check
      * for your .env file needs to know which keys need to be present.