+++ src/Checks/SchedulerHealthCheck.php
@@ -8,25 +8,26 @@ use Illuminate\Support\Facades\Storage;
 class SchedulerHealthCheck extends HealthCheck
 {
     protected $name = 'scheduler';
-
-    const FILE_NAME = 'laravel-scheduler-health-check.txt';
     
     public function status()
     {
-        if (!Storage::exists(static::FILE_NAME)) {
+        $filename = config('healthcheck.scheduler.timestamp-filename');
+        $minutesBetweenChecks = config('healthcheck.scheduler.minutes-between-checks');
+
+        if (!Storage::exists($filename)) {
             return $this->problem('Scheduler has not ran yet');
         }
 
-        $schedulerLastRan = Storage::get(static::FILE_NAME);
+        $schedulerLastRan = Storage::get($filename);
         $now = time();
 
         $secondsSinceSchedulerRan = $now - $schedulerLastRan;
 
-        if ($secondsSinceSchedulerRan < 60) {
+        if ($secondsSinceSchedulerRan < (60 * $minutesBetweenChecks)) {
             return $this->okay();
         }
 
-        return $this->problem('Scheduler last ran more than 1 minute ago', [
+        return $this->problem("Scheduler last ran more than $minutesBetweenChecks minute ago", [
             'scheduler_last_ran' => $schedulerLastRan,
             'now' => $now,
             'time_since_scheduler_ran' => $secondsSinceSchedulerRan,